Move Select logic to data_interface and remove ui module
